functions/fms_create_postgresql_publication_replication/requirements.txt
pg8000==1.29.8
SQLAlchemy==2.0.18
cloud-sql-python-connector==1.2.4
google-cloud-secret-manager==2.12.6
google-cloud-logging==3.4.0